updated-bot/requirements.txt
requests
beautifulsoup4
pdfplumber
openai
python-telegram-bot==20.0
python-dotenv
